pyproject.toml
[project]
name = "sentry-django-demo"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "sentry-sdk>=2.24",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.7",
    "python-dotenv>=1.0",
]
